+++ shell/src/main/java/com/samsung/command/support/ArgumentValidator.java
@@ -1,4 +1,4 @@
-package com.samsung.validator;
+package com.samsung.command.support;
 
 public class ArgumentValidator {
