+++ shell/src/main/java/com/samsung/command/support/CommandValidator.java
@@ -1,4 +1,4 @@
-package com.samsung.validator;
+package com.samsung.command.support;
 
 public class CommandValidator {
