+++ shell/src/test/java/com/samsung/command/support/CommandValidatorTest.java
@@ -1,4 +1,4 @@
-package com.samsung.validator;
+package com.samsung.command.support;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
